+++ CONTRIBUTING.md
@@ -1,0 +1,6 @@
+# Contribution guidelines
+
+## Workflow
+
+* Develop on a feature branch and send a pull request for review.
+* Assign the pull request to Matt Watson [@mwtsn](https://github.com/@mwtsn).
